Accommodate code-review and simplify the NomadConfig class structure [ci skip]
