+++ plugins/nf-nomad/src/main/nextflow/nomad/config/VolumeSpec.groovy
@@ -1,5 +1,26 @@
-package nextflow.nomad.config
+/*
+ * Copyright 2023-, Stellenbosch University, South Africa
+ * Copyright 2024, Evaluacion y Desarrollo de Negocios, Spain
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
+package nextflow.nomad.config
+/**
+ * Nomad Volume Spec
+ *
+ * @author Jorge Aguilera <jagedn@gmail.com>
+ */
 class VolumeSpec {
 
     final static public String VOLUME_DOCKER_TYPE = "docker"
